requirements.txt
Flask==3.0.2
psycopg[binary]==3.3.3
python-dotenv==1.0.1
